+++ src/main/kotlin/blackjack/view/model/PlayerConfig.kt
@@ -2,5 +2,5 @@ package blackjack.view.model
 
 class PlayerConfig(
     val name: String,
-    val bettingAmount: Int,
+    val bettingAmount: Double,
 )
